requirements.txt
torch
torchvision
timm==0.4.12
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
deepspeed==0.4.0
scipy
ase==3.21.1
braceexpand==0.1.7
cached-property==1.5.2
configparser==5.0.2
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
gitdb==4.0.7
gitpython==3.1.14
h5py==3.1.0
isodate==0.6.0
jinja2==3.0.1
kiwisolver==1.3.1
littleutils==0.2.2
llvmlite==0.36.0
markupsafe==2.0.1
matplotlib==3.3.4
networkx==2.5.1
numba==0.53.1
ogb==1.3.1
outdated==0.2.1
pathtools==0.1.2
promise==2.3
psutil==5.8.0
pyarrow==4.0.0
pyparsing==2.4.7
python-louvain==0.15
pyyaml==5.4.1
rdflib==5.0.0
sentry-sdk==1.1.0
shortuuid==1.0.1
sklearn==0.0
smmap==4.0.0
subprocess32==3.5.4
torch-geometric==1.7.0
wandb==0.10.30
wilds==1.1.0
google-cloud-storage
protobuf==3.17.3
torchsummary